Refactor: Extract Playwright script helpers to separate file
    Moved helper functions (replace_sensitive_data, _try_locate_and_act, PlaywrightActionError) from a multiline string within PlaywrightScriptGenerator into browser_use/agent/playwright_script_helpers.py.

    This allows linters and type checkers to analyze the helper code, improving maintainability and code quality. The generator now reads the helper file content dynamically.
